Split STR upload page
