+++ src/applications/disability-benefits/all-claims/config/form.js
@@ -37,6 +37,7 @@ import {
   directToCorrectForm,
   DISABILITY_SHARED_CONFIG,
   isBDD,
+  isUploadingSTR,
   showSeparationLocation,
   getPageTitle,
   claimingNew,
@@ -91,6 +92,7 @@ import {
   separationPay,
   servedInCombatZone,
   serviceTreatmentRecords,
+  serviceTreatmentRecordsAttachments,
   socialBehaviorChanges,
   summaryOfDisabilities,
   summaryOfEvidence,
@@ -584,6 +586,13 @@ const formConfig = {
           depends: formData => isBDD(formData),
           uiSchema: serviceTreatmentRecords.uiSchema,
           schema: serviceTreatmentRecords.schema,
+        },
+        serviceTreatmentRecordsAttachments: {
+          title: 'Service treatment records upload',
+          path: 'supporting-evidence/service-treatment-records-upload',
+          depends: formData => isUploadingSTR(formData),
+          uiSchema: serviceTreatmentRecordsAttachments.uiSchema,
+          schema: serviceTreatmentRecordsAttachments.schema,
         },
         evidenceTypes: {
           title: 'Supporting evidence types',
