+++ src/applications/disability-benefits/all-claims/pages/serviceTreatmentRecords.js
@@ -1,22 +1,6 @@
 import _ from 'platform/utilities/data';
-import { validateFileField } from 'platform/forms-system/src/js/validation';
 
-import { ancillaryFormUploadUi } from '../utils';
-import { UploadDescription } from '../content/fileUploadDescriptions';
 import { serviceTreatmentRecordsSubmitLater } from '../content/serviceTreatmentRecords';
-// import fullSchema from 'vets-json-schema/dist/21-526EZ-ALLCLAIMS-schema.json';
-import fullSchema from '../config/schema';
-
-const fileUploadUi = ancillaryFormUploadUi(
-  'Upload your service treatment records',
-  '',
-  {
-    attachmentId: '',
-    addAnotherLabel: 'Add another document',
-  },
-);
-
-const { serviceTreatmentRecordsAttachments } = fullSchema.properties;
 
 export const uiSchema = {
   'view:uploadServiceTreatmentRecordsQualifier': {
@@ -38,23 +22,6 @@ export const uiSchema = {
         useDlWrap: true,
       },
     },
-  },
-  serviceTreatmentRecordsAttachments: {
-    ...fileUploadUi,
-    'ui:options': {
-      ...fileUploadUi['ui:options'],
-      expandUnder: 'view:uploadServiceTreatmentRecordsQualifier',
-      expandUnderCondition: data =>
-        _.get('view:hasServiceTreatmentRecordsToUpload', data, false),
-    },
-    'ui:description': UploadDescription,
-    'ui:required': data =>
-      _.get(
-        'view:uploadServiceTreatmentRecordsQualifier.view:hasServiceTreatmentRecordsToUpload',
-        data,
-        false,
-      ),
-    'ui:validations': [validateFileField],
   },
   'view:serviceTreatmentRecordsSubmitLater': {
     'ui:title': '',
@@ -82,7 +49,6 @@ export const schema = {
         },
       },
     },
-    serviceTreatmentRecordsAttachments,
     'view:serviceTreatmentRecordsSubmitLater': {
       type: 'object',
       properties: {},
